Move calculation ingest code into calculation endpoint
