+++ girder/molecules/server/molecule.py
@@ -171,53 +171,6 @@ class Molecule(Resource):
                     semantic.upload_molecule(mol)
                 except requests.ConnectionError:
                     print(TerminalColor.warning('WARNING: Couldn\'t connect to virtuoso.'))
-
-            if 'vibrations' in cjson or 'basisSet' in cjson:
-                # We have some calculation data, let's add it to the calcs.
-                sdf = output
-                moleculeId = mol['_id']
-                calc_props = {}
-
-                if calc_id is not None:
-                    calc = self._calc_model.load(calc_id, user=user, level=AccessType.ADMIN)
-                    calc_props = calc['properties']
-                    # The calculation is no longer pending
-                    if 'pending' in calc_props:
-                        del calc_props['pending']
-
-                if input_format == 'json':
-                    jsonInput = json.loads(data_str)
-                    # Don't override existing properties
-                    new_calc_props = avogadro.calculation_properties(jsonInput)
-                    new_calc_props.update(calc_props)
-                    calc_props = new_calc_props
-
-                # Use basisSet from cjson if we don't already have one.
-                if 'basisSet' in cjson and 'basisSet' not in calc_props:
-                    calc_props['basisSet'] = cjson['basisSet']
-
-                # Use functional from cjson properties if we don't already have
-                # one.
-                functional = parse('properties.functional').find(cjson)
-                if functional and 'functional' not in calc_props:
-                    calc_props['functional'] = functional[0].value
-
-                # Add theory priority to 'sort' calculations
-                theory = calc_props.get('theory')
-                functional = calc_props.get('functional')
-                if theory in constants.theory_priority:
-                    priority = constants.theory_priority[theory]
-                    calc_props['theoryPriority'] = priority
-
-                if calc_id is not None:
-                    calc['properties'] = calc_props
-                    calc['cjson'] = cjson
-                    calc['fileId'] = file_id
-                    self._calc_model.save(calc)
-                else:
-                    self._calc_model.create_cjson(user, cjson, calc_props,
-                                                  moleculeId, file_id, public)
-
         elif 'xyz' in body or 'sdf' in body:
 
             if 'xyz' in body:
